Removed OIDC login flow until it will be relevant again
