--- cucumber/authenticators/features/authn_oidc_login.feature_disable
@@ -1,54 +1,0 @@
-In order not to delete our previous OIDC authenticator development which stopped in the middle due to change in priority,
-We decide to mark this feature as shutdown.
-In addition we want to use the same url of the authenticate API for this new feature which is also an OIDC authenticator,
-But for different use case.
-
-Feature: Users can authneticate with OIDC authenticator
-
-  Background:
-    Given a policy:
-    """
-    - !user alice
-
-    - !policy
-      id: conjur/authn-oidc/keycloak
-      body:
-      - !webservice
-        annotations:
-          description: Authentication service for Keycloak, based on Open ID Connect.
-
-      - !variable
-        id: client-id
-
-      - !variable
-        id: client-secret
-
-      - !variable
-        id: provider-uri
-
-      - !variable
-        id: id-token-user-property
-
-      - !group users
-
-      - !permit
-        role: !group users
-        privilege: [ read, authenticate ]
-        resource: !webservice
-
-    - !grant
-      role: !group conjur/authn-oidc/keycloak/users
-      member: !user alice
-    """
-
-    And I am the super-user
-    And I successfully set OIDC variables
-
-
-  Scenario: A valid authorization code and redirect uri to get Conjur access token
-    Given I get authorization code
-    When I successfully login via OIDC
-    Then login response token is valid
-
-    When I successfully authenticate via OIDC
-    Then user "alice" is authorized
